pyproject.toml
[project]
name = "rag-application"
version = "0.1.0"
description = "A RAG (Retrieval-Augmented Generation) application"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "python-dotenv>=1.0.0",
    "pdfplumber==0.10.3",
    "pdfminer.six==20221105",
    "llama-cpp-python>=0.2.6"
]
authors = [
    {name = "Your Name"}
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]